refactor: 보드 수정중 참여자 초대 modal 컴포넌트 리팩토링
